[DAGE-122] refactored escape into BaseSearchEngine
