+++ rre-tools/tests/shared/search_engines/test_solr_search_engine.py
@@ -127,27 +127,3 @@ def test_solr_search_engine_negative_post_fetch_for_evaluation__expects__raises_
 def test_solr_search_engine_bad_url__expects__raises_validation_error():
     with pytest.raises(ValidationError):
         _ = SolrSearchEngine("fake-NONurl")
-
-
-def test_escape_no_special_chars():
-    assert SolrSearchEngine.escape("hello world") == "hello world"
-
-def test_escape_basic_specials():
-    assert SolrSearchEngine.escape("a+b") == "a\\+b"
-    assert SolrSearchEngine.escape("field:value") == "field\\:value"
-    assert SolrSearchEngine.escape("(test)") == "\\(test\\)"
-
-def test_escape_multiple_specials():
-    assert SolrSearchEngine.escape("a+b-(c*d)?") == "a\\+b\\-\\(c\\*d\\)\\?"
-    assert SolrSearchEngine.escape("[range]~{json}") == "\\[range\\]\\~\\{json\\}"
-
-def test_escape_with_backslash():
-    assert SolrSearchEngine.escape("path\\to/file") == "path\\\\to\\/file"
-
-def test_escape_quotes():
-    assert SolrSearchEngine.escape('"phrase" AND title:book') == '\\"phrase\\" AND title\\:book'
-
-def test_escape_all_special_chars():
-    all_specials = r'\+-!():^[]"{}~*?|&/'
-    expected = ''.join(['\\' + c for c in all_specials])
-    assert SolrSearchEngine.escape(all_specials) == expected
